refactor(external_requests): rename get_soup.py module to more accurate and descriptive external_requests.py
