+++ src/scraper/external_requests.py
