+++ src/scraper/sites/fed_new_york_scraper.py
@@ -1,5 +1,5 @@
 from ..generic_scraper import GenericScraper
-from src.scraper.get_soup import request_json, request_soup
+from src.scraper.external_requests import request_json, request_soup
 from datetime import datetime
 import requests
 import warnings
