+++ src/scraper/sites/fed_san_francisco_scraper.py
@@ -1,5 +1,5 @@
 from ..generic_scraper import GenericScraper
-from src.scraper.get_soup import request_json
+from src.scraper.external_requests import request_json
 from bs4 import BeautifulSoup
 
 class FedSanFranciscoScraper(GenericScraper):
